combine title and og tags in PageMeta component
